pyproject.toml
[project]
name = "snapctl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.37",
    "pyyaml>=6.0.2",
    "ruff>=0.13.1",
    "rich>=14.1.0",
    "typer>=0.19.2",
    "structlog>=25.5.0",
    "pytest>=9.0.1",
    "pytest-cov>=6.0.0",
    "mypy>=1.0",
    "boto3-stubs>=1.34",
    "types-PyYAML>=6.0",
    "moto>=5.0",
    "bandit>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli"]

[project.scripts]
snapctl = "cli.commands.main:app"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
# Start with lenient settings
disallow_untyped_defs = false
ignore_missing_imports = true

# Exclude test files for now
exclude = [
    "cli/tests/.*",
]
